[JetNews] Move AppDrawer Composable from JetnewsApp.kt to AppDrawer.kt
