+++ JetNews/app/src/main/java/com/example/jetnews/ui/JetnewsApp.kt
@@ -17,34 +17,9 @@
 package com.example.jetnews.ui
 
 import androidx.compose.animation.Crossfade
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.preferredHeight
-import androidx.compose.foundation.layout.preferredWidth
-import androidx.compose.material.Divider
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.material.TextButton
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.ListAlt
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.graphics.vector.VectorAsset
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.unit.dp
-import androidx.ui.tooling.preview.Preview
-import com.example.jetnews.R
 import com.example.jetnews.data.AppContainer
 import com.example.jetnews.data.interests.InterestsRepository
 import com.example.jetnews.data.posts.PostsRepository
@@ -91,134 +66,5 @@ private fun AppContent(
                 )
             }
         }
-    }
-}
-
-@Composable
-fun AppDrawer(
-    navigateTo: (Screen) -> Unit,
-    currentScreen: Screen,
-    closeDrawer: () -> Unit
-) {
-    Column(modifier = Modifier.fillMaxSize()) {
-        Spacer(Modifier.preferredHeight(24.dp))
-        JetNewsLogo(Modifier.padding(16.dp))
-        Divider(color = MaterialTheme.colors.onSurface.copy(alpha = .2f))
-        DrawerButton(
-            icon = Icons.Filled.Home,
-            label = "Home",
-            isSelected = currentScreen == Screen.Home,
-            action = {
-                navigateTo(Screen.Home)
-                closeDrawer()
-            }
-        )
-
-        DrawerButton(
-            icon = Icons.Filled.ListAlt,
-            label = "Interests",
-            isSelected = currentScreen == Screen.Interests,
-            action = {
-                navigateTo(Screen.Interests)
-                closeDrawer()
-            }
-        )
-    }
-}
-
-@Composable
-private fun JetNewsLogo(modifier: Modifier = Modifier) {
-    Row(modifier = modifier) {
-        Image(
-            asset = vectorResource(R.drawable.ic_jetnews_logo),
-            colorFilter = ColorFilter.tint(MaterialTheme.colors.primary)
-        )
-        Spacer(Modifier.preferredWidth(8.dp))
-        Image(
-            asset = vectorResource(R.drawable.ic_jetnews_wordmark),
-            colorFilter = ColorFilter.tint(MaterialTheme.colors.onSurface)
-        )
-    }
-}
-
-@Composable
-private fun DrawerButton(
-    icon: VectorAsset,
-    label: String,
-    isSelected: Boolean,
-    action: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val colors = MaterialTheme.colors
-    val imageAlpha = if (isSelected) {
-        1f
-    } else {
-        0.6f
-    }
-    val textIconColor = if (isSelected) {
-        colors.primary
-    } else {
-        colors.onSurface.copy(alpha = 0.6f)
-    }
-    val backgroundColor = if (isSelected) {
-        colors.primary.copy(alpha = 0.12f)
-    } else {
-        Color.Transparent
-    }
-
-    val surfaceModifier = modifier
-        .padding(start = 8.dp, top = 8.dp, end = 8.dp)
-        .fillMaxWidth()
-    Surface(
-        modifier = surfaceModifier,
-        color = backgroundColor,
-        shape = MaterialTheme.shapes.small
-    ) {
-        TextButton(
-            onClick = action,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Row(
-                horizontalArrangement = Arrangement.Start,
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Image(
-                    asset = icon,
-                    colorFilter = ColorFilter.tint(textIconColor),
-                    alpha = imageAlpha
-                )
-                Spacer(Modifier.preferredWidth(16.dp))
-                Text(
-                    text = label,
-                    style = MaterialTheme.typography.body2,
-                    color = textIconColor
-                )
-            }
-        }
-    }
-}
-
-@Preview("Drawer contents")
-@Composable
-fun PreviewJetnewsApp() {
-    ThemedPreview {
-        AppDrawer(
-            navigateTo = { },
-            currentScreen = Screen.Home,
-            closeDrawer = { }
-        )
-    }
-}
-
-@Preview("Drawer contents dark theme")
-@Composable
-fun PreviewJetnewsAppDark() {
-    ThemedPreview(darkTheme = true) {
-        AppDrawer(
-            navigateTo = { },
-            currentScreen = Screen.Home,
-            closeDrawer = { }
-        )
     }
 }
